requirements.txt
requests
beautifulsoup4
openai
python-telegram-bot